Delete unused files and add cleaned up files
